+++ src-tauri/src/constants.rs
@@ -13,7 +13,7 @@ pub mod network {
         pub const DEFAULT_HTTP: u16 = 7899;
 
         #[cfg(not(feature = "celestial-dev"))]
-        pub const SINGLETON_SERVER: u16 = 33331;
+        pub const SINGLETON_SERVER: u16 = 33341;
         #[cfg(feature = "celestial-dev")]
         pub const SINGLETON_SERVER: u16 = 11233;
     }